Refactor: separate SectionTabList component button
- SectionTabList 컴포넌트의 버튼을 SectionTabListButton 컴포넌트로 분리
- handleTabClick 함수명과 매개변수명 변경
