+++ src/components/SectionTabs/SectionTabList.jsx
@@ -1,11 +1,8 @@
 import { useRef } from 'react';
-import { useTheme, css } from '@emotion/react';
-import {
-  useActiveSectionTab,
-  useSetActiveSectionTab,
-} from '@contexts/ActiveSectionTabContext';
+import { css } from '@emotion/react';
 import useTouchScroll from '@hooks/useTouchScroll';
 import sectionTabList from '@constants/sectionTabList';
+import { SectionTabListButton } from '@components/SectionTabs';
 import { mediaQuery } from '@styles/breakpoints';
 
 function SectionTabList() {
@@ -19,14 +16,6 @@ function SectionTabList() {
     handleScrollMove,
     handleScrollEnd,
   ] = useTouchScroll(scrollRef);
-
-  const activeSectionTab = useActiveSectionTab();
-  const setActiveSectionTab = useSetActiveSectionTab();
-  const theme = useTheme();
-
-  const handleTabClick = (tab) => {
-    if (scrollStartClientX == scrollEndClientX) setActiveSectionTab(tab);
-  };
 
   return (
     <ul
@@ -57,42 +46,12 @@ function SectionTabList() {
             position: relative;
           `}
         >
-          <button
-            onClick={() => handleTabClick(sectionTab)}
-            css={css`
-              height: 3rem;
-              padding: 0 1.5rem;
-              color: ${sectionTab === activeSectionTab
-                ? theme.color.emphasis.primary
-                : theme.color.text.disabled};
-              cursor: ${isScrolling ? 'grabbing' : null};
-              transition: all 0.2s;
-              ${theme.typography.label1}
-              ::after {
-                content: '';
-                width: 100%;
-                height: 3px;
-                border-radius: 2px;
-                background-color: ${theme.color.emphasis.primary};
-                position: absolute;
-                left: 0;
-                bottom: 0;
-                transform: scale(${sectionTab === activeSectionTab ? 1 : 0});
-                transition: all 0.2s;
-              }
-              :hover {
-                color: ${theme.color.emphasis.primary};
-              }
-              ${mediaQuery[2]} {
-                padding: 0 1.25rem;
-              }
-              ${mediaQuery[1]} {
-                padding: 0 1rem;
-              }
-            `}
-          >
-            {sectionTab}
-          </button>
+          <SectionTabListButton
+            sectionTab={sectionTab}
+            isScrolling={isScrolling}
+            scrollStartClientX={scrollStartClientX}
+            scrollEndClientX={scrollEndClientX}
+          />
         </li>
       ))}
     </ul>
