+++ src/components/SectionTabs/SectionTabListButton.jsx
@@ -1,0 +1,64 @@
+import { useTheme, css } from '@emotion/react';
+import {
+  useActiveSectionTab,
+  useSetActiveSectionTab,
+} from '@contexts/ActiveSectionTabContext';
+import { mediaQuery } from '@styles/breakpoints';
+
+function SectionTabListButton({
+  sectionTab,
+  isScrolling,
+  scrollStartClientX,
+  scrollEndClientX,
+}) {
+  const activeSectionTab = useActiveSectionTab();
+  const setActiveSectionTab = useSetActiveSectionTab();
+  const theme = useTheme();
+
+  const handleTabListButtonClick = (tabListButton) => {
+    if (scrollStartClientX == scrollEndClientX)
+      setActiveSectionTab(tabListButton);
+  };
+
+  return (
+    <button
+      onClick={() => handleTabListButtonClick(sectionTab)}
+      css={css`
+        height: 3rem;
+        padding: 0 1.5rem;
+        color: ${sectionTab === activeSectionTab
+          ? theme.color.emphasis.primary
+          : theme.color.text.disabled};
+        cursor: ${isScrolling ? 'grabbing' : null};
+        transition: all 0.2s;
+        ${theme.typography.label1}
+        ::after {
+          content: '';
+          width: 100%;
+          height: 3px;
+          border-radius: 2px;
+          background-color: ${theme.color.emphasis.primary};
+          position: absolute;
+          left: 0;
+          bottom: 0;
+          transform: scale(${sectionTab === activeSectionTab ? 1 : 0});
+          transition: all 0.2s;
+        }
+        :hover {
+          color: ${theme.color.emphasis.primary};
+        }
+
+        ${mediaQuery[2]} {
+          padding: 0 1.25rem;
+        }
+        ${mediaQuery[1]} {
+          padding: 0 1rem;
+        }
+      `}
+    >
+      {sectionTab}
+    </button>
+  );
+}
+
+export default SectionTabListButton;
