+++ src/components/SectionTabs/index.js
@@ -1,6 +1,13 @@
 import SectionTabList from './SectionTabList';
+import SectionTabListButton from './SectionTabListButton';
 import ArticleList from './ArticleList';
 import Article from './Article';
 import ViewMoreButton from './ViewMoreButton';
 
-export { SectionTabList, ArticleList, Article, ViewMoreButton };
+export {
+  SectionTabList,
+  SectionTabListButton,
+  ArticleList,
+  Article,
+  ViewMoreButton,
+};
